Made a skill table and switched to css
I don't know how to change the text color yet, but I think now that I went to css it'll be easier. The table works and can be included in a dropdown section thingy. So now I just need to make the skilltable component more modular so that it displays any section of skills instead of the hardcoded set of social skills.